// src/models/redux/enums.ts
export enum AdminActionEnum {
  START_LOADING = 'START_LOADING',
  ADMIN_LOGIN = 'ADMIN_LOGIN',
  ADMIN_LOGOUT = 'ADMIN_LOGOUT',
  ADMIN_ERROR_OPEN_SNACKBAR = 'ADMIN_ERROR_OPEN_SNACKBAR',
  ADMIN_AUTHORIZED_ERROR = 'ADMIN_AUTHORIZED_ERROR',
  ADMIN_AUTHORIZED_CHECK = 'ADMIN_AUTHORIZED_CHECK',
}

export enum CategoryActionEnum {
  GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES',
  START_LOADING = 'START_LOADING',
  GET_PART_OF_CATEGORIES = 'GET_PART_OF_CATEGORIES',
  GET_ALL_AVAILABLE_CAEGORIES = 'GET_ALL_AVAILABLE_CAEGORIES',
  CREATE_CATEGORY = 'CREATE_CATEGORY',
  UPDATE_CATEGORY = 'UPDATE_CATEGORY',
  CATEGORIES_ERRORS = 'CATEGORIES_ERRORS',
  CATEGORIES_ERROR_CLOSE_SNACKBAR = 'USER_ERROR_CLOSE_SNACKBAR',
  FIND_CATEGORY = 'FIND_CATEGORY',
}

export enum ProductsActionEnum {
  GET_ALL_PRODUCTS = 'GET_ALL_PRODUCTS',
  START_LOADING = 'START_LOADING',
  GET_PRODUCTS_BY_CATEGORY = 'GET_PRODUCTS_BY_CATEGORY',
  GET_PART_OF_PRODUCTS = 'GET_PART_OF_PRODUCTS',
  GET_AVAILABLE_PRODUCTS = 'GET_AVAILABLE_PRODUCTS',
  HIDE_PRODUCT = 'HIDE_PRODUCT',
  SHOW_PRODUCT = 'SHOW_PRODUCT',
  FIND_PRODUCT = 'FIND_PRODUCT',
  CREATE_PRODUCT = 'CREATE_PRODUCT',
  UPDATE_PRODUCT = 'UPDATE_PRODUCT',
  PRODUCTS_ERRORS = 'PRODUCTS_ERRORS',
  PRODUCTS_ERROR_CLOSE_SNACKBAR = 'USER_ERROR_CLOSE_SNACKBAR',
}

export enum PickUpPointActionEnum {
  GET_PART_OF_PICK_UP_POINTS = 'GET_PART_OF_PICK_UP_POINTS',
  FIND_PICK_UP_POINTS = 'FIND_PICK_UP_POINTS',
  CREATE_PICK_UP_POINTS = 'CREATE_PICK_UP_POINTS',
  UPDATE_PICK_UP_POINTS = 'UPDATE_PICK_UP_POINTS',
  HIDE_PICK_UP_POINTS = 'HIDE_PICK_UP_POINTS',
  SHOW_PICK_UP_POINTS = 'SHOW_PICK_UP_POINTS',
}